chore: ServiceStatus builder methods
