+++ bin/network-monitor/src/note_transport.rs
@@ -11,8 +11,8 @@ use tonic_health::pb::{HealthCheckRequest, health_check_response};
 use tracing::{info, instrument};
 use url::Url;
 
-use crate::status::{NoteTransportStatusDetails, ServiceDetails, ServiceStatus, Status};
-use crate::{COMPONENT, current_unix_timestamp_secs};
+use crate::COMPONENT;
+use crate::status::{NoteTransportStatusDetails, ServiceDetails, ServiceStatus};
 
 /// Creates a `tonic` channel for the given URL, enabling TLS for `https` schemes.
 fn create_channel(url: &Url, timeout: Duration) -> Result<Channel, tonic::transport::Error> {
@@ -42,15 +42,8 @@ pub async fn run_note_transport_status_task(
     loop {
         interval.tick().await;
 
-        let current_time = current_unix_timestamp_secs();
-
-        let status = check_note_transport_status(
-            &mut health_client,
-            url.to_string(),
-            name.clone(),
-            current_time,
-        )
-        .await;
+        let status =
+            check_note_transport_status(&mut health_client, url.to_string(), name.clone()).await;
 
         if status_sender.send(status).is_err() {
             info!("No receivers for note transport status updates, shutting down");
@@ -70,7 +63,6 @@ pub(crate) async fn check_note_transport_status(
     health_client: &mut HealthClient<Channel>,
     url: String,
     name: String,
-    current_time: u64,
 ) -> ServiceStatus {
     let request = HealthCheckRequest { service: String::new() };
 
@@ -78,42 +70,30 @@ pub(crate) async fn check_note_transport_status(
         Ok(response) => {
             let serving_status = response.into_inner().status();
             let is_serving = serving_status == health_check_response::ServingStatus::Serving;
-
-            let status = if is_serving { Status::Healthy } else { Status::Unhealthy };
             let serving_status_str = format!("{serving_status:?}");
 
-            ServiceStatus {
-                name,
-                status,
-                last_checked: current_time,
-                error: None,
-                details: ServiceDetails::NoteTransportStatus(NoteTransportStatusDetails {
+            if is_serving {
+                let details = ServiceDetails::NoteTransportStatus(NoteTransportStatusDetails {
                     url,
                     serving_status: serving_status_str,
-                }),
+                });
+                ServiceStatus::healthy(name, details)
+            } else {
+                let error = format!("serving status: {serving_status_str}");
+                let details = ServiceDetails::NoteTransportStatus(NoteTransportStatusDetails {
+                    url,
+                    serving_status: serving_status_str,
+                });
+                ServiceStatus::unhealthy(name, error, details)
             }
         },
-        Err(e) => unhealthy(&name, current_time, &e),
-    }
-}
-
-/// Returns an unhealthy service status.
-fn unhealthy(name: &str, current_time: u64, err: &impl ToString) -> ServiceStatus {
-    ServiceStatus {
-        name: name.to_owned(),
-        status: Status::Unhealthy,
-        last_checked: current_time,
-        error: Some(err.to_string()),
-        details: ServiceDetails::Error,
+        Err(e) => ServiceStatus::error(name, e),
     }
 }
 
 pub(crate) fn initial_note_transport_status() -> ServiceStatus {
-    ServiceStatus {
-        name: "Note Transport".to_string(),
-        status: Status::Unknown,
-        last_checked: current_unix_timestamp_secs(),
-        error: None,
-        details: ServiceDetails::NoteTransportStatus(NoteTransportStatusDetails::default()),
-    }
+    ServiceStatus::unknown(
+        "Note Transport",
+        ServiceDetails::NoteTransportStatus(NoteTransportStatusDetails::default()),
+    )
 }
